+++ packages/app-config-writer/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @sap-ux/app-config-writer
 
+## 0.3.17
+
+### Patch Changes
+
+-   Updated dependencies [e6f454ab]
+    -   @sap-ux/project-access@1.16.3
+    -   @sap-ux/axios-extension@1.8.1
+
+## 0.3.16
+
+### Patch Changes
+
+-   Updated dependencies [3000e8f4]
+    -   @sap-ux/axios-extension@1.8.1
+
+## 0.3.15
+
+### Patch Changes
+
+-   Updated dependencies [62232236]
+    -   @sap-ux/axios-extension@1.8.0
+
+## 0.3.14
+
+### Patch Changes
+
+-   Updated dependencies [d5b99a28]
+    -   @sap-ux/project-access@1.16.2
+    -   @sap-ux/axios-extension@1.7.3
+
 ## 0.3.13
 
 ### Patch Changes