+++ packages/backend-proxy-middleware/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @sap-ux/backend-proxy-middleware
 
+## 0.7.8
+
+### Patch Changes
+
+-   10ad4b51: Use token in backend-proxy-middleware when connecting to abap cloud
+
+## 0.7.7
+
+### Patch Changes
+
+-   Updated dependencies [3000e8f4]
+    -   @sap-ux/axios-extension@1.8.1
+
+## 0.7.6
+
+### Patch Changes
+
+-   Updated dependencies [62232236]
+    -   @sap-ux/axios-extension@1.8.0
+
 ## 0.7.5
 
 ### Patch Changes