+++ packages/project-access/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/project-access
 
+## 1.16.3
+
+### Patch Changes
+
+-   e6f454ab: Fix for detecting apps that are part of CAP
+
+## 1.16.2
+
+### Patch Changes
+
+-   d5b99a28: cds.root is now set from passed project root
+
 ## 1.16.1
 
 ### Patch Changes