+++ packages/ui5-library-writer/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sap-ux/ui5-library-writer
 
+## 0.3.10
+
+### Patch Changes
+
+-   Updated dependencies [e6f454ab]
+    -   @sap-ux/project-access@1.16.3
+    -   @sap-ux/ui5-application-writer@0.26.2
+
+## 0.3.9
+
+### Patch Changes
+
+-   Updated dependencies [d5b99a28]
+    -   @sap-ux/project-access@1.16.2
+    -   @sap-ux/ui5-application-writer@0.26.2
+
 ## 0.3.8
 
 ### Patch Changes